+++ src/main/java/com/control/web_market/removeBag.java
@@ -1,5 +1,6 @@
-package com.model.web_market;
+package com.control.web_market;
 
+import com.control.web_market.Product;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
