+++ src/main/java/com/model/web_market/MainPage.java
@@ -1,5 +1,6 @@
 package com.model.web_market;
 
+import com.control.web_market.Product;
 import jakarta.servlet.ServletException;
 import java.io.*;
 import jakarta.servlet.http.*;
